Moved URLCamera device to its own file
